Refactor tests module to a subpackage
