+++ gaeauth/tests/__init__.py
@@ -1,0 +1,2 @@
+from gaeauth.tests.backends import *
+from gaeauth.tests.models import *
