+++ gaeauth/tests/backends.py
@@ -7,16 +7,6 @@ from gaeauth.backends import GoogleAccountBackend
 from google.appengine.api import users
 
 
-class ModelsTest(TestCase):
-    def test_password_indexed(self):
-        """Tests that the password field on the User model is indexed."""
-        password_indexed = False
-        for field in User._meta.local_fields:
-            if field.column == 'password':
-                password_indexed = field.db_index
-        self.assertTrue(password_indexed)
-
-
 class GoogleAccountBackendTest(TestCase):
     def setUp(self):
         settings.AUTHENTICATION_BACKENDS = (
